Change component/instance maps to `PrimaryMap` (#10916)
* Change component/instance maps to `PrimaryMap`

This switches to using typed keys for these maps to be more idiomatic
with the rest of Wasmtime and this has the additional benefit of
compressing indices to 32-bits instead of the previous
pointer-sized-bits.

* Fix an unused import

+++ crates/wasmtime/src/runtime/component/instance.rs
@@ -42,7 +42,7 @@ pub struct Instance {
 // in-Rust representation here in terms of size/alignment/etc.
 const _: () = {
     #[repr(C)]
-    struct C(u64, usize);
+    struct C(u64, u32);
     assert!(core::mem::size_of::<C>() == core::mem::size_of::<Instance>());
     assert!(core::mem::align_of::<C>() == core::mem::align_of::<Instance>());
     assert!(core::mem::offset_of!(Instance, id) == 0);
